BAH-4552 | Refactor. Rename ImmunizationHistory Form As Immunization Form
