+++ apps/clinical/src/components/forms/immunization/__tests__/SelectedImmunizationItem.test.tsx
