+++ apps/clinical/src/components/forms/immunization/components/SelectedImmunizationItem.tsx
